+++ patches/std_kernel_stub.h
@@ -1,9 +1,6 @@
 #ifndef _std_kernel_h
 #define _std_kernel_h
 
-// Minimal stub for rsta2/circle build (headless, no display)
-// Replaces phorton1 std_kernel.h to avoid CScreenDeviceBase/ws/ dependency
-
 #define USE_UI_SYSTEM		0
 #define USE_AUDIO_SYSTEM	1
 #define USE_MIDI_SYSTEM		0
@@ -15,7 +12,18 @@
 #define USE_ILI_TFT		0
 #define USE_XPT2046		0
 
+#ifdef ARM_ALLOW_MULTI_CORE
+#define CORE_FOR_AUDIO_SYSTEM	1
+#define IPI_AUDIO_UPDATE	11
+class CCoreTask {
+public:
+	static CCoreTask *Get();
+	void SendIPI(unsigned nCore, unsigned nIPI);
+};
+#else
 #define CORE_FOR_AUDIO_SYSTEM	0
+#endif
+
 #define CORE_FOR_UI_SYSTEM	2
 
 #include <circle/types.h>